Move Angular Client Core into a module
