+++ modules/core/common/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@module/core-common",
+  "version": "1.0.0",
+  "private": true
+}
