+++ modules/core/common/typings.d.ts
@@ -1,0 +1,2 @@
+declare var __SERVER_PORT__: number;
+declare var __API_URL__: string;
