+++ modules/core/server-ts/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@module/core-server-ts",
+  "version": "1.0.0",
+  "private": true,
+  "dependencies": {
+    "@module/core-common": "^1.0.0"
+  }
+}
+
