+++ package.json
@@ -159,7 +159,7 @@
   },
   "dependencies": {
     "opencollective": "^1.0.3",
-    "spinjs": "^0.4.175"
+    "spinjs": "^0.4.176"
   },
   "resolutions": {
     "react": "16.5.0",
