+++ packages/client-angular/package.json
@@ -45,6 +45,7 @@
     "@angular/platform-browser-dynamic": "^7.0.3",
     "@angular/router": "^7.0.3",
     "@angularclass/hmr": "^2.1.3",
+    "@module/core-client-angular": "^1.0.0",
     "@module/counter-client-angular": "^1.0.0",
     "@module/counter-common": "^1.0.0",
     "@module/module-client-angular": "^1.0.0",
