--- packages/client-angular/src/modules/index.ts
@@ -1,7 +1,0 @@
-import counter from '@module/counter-client-angular';
-import ClientModule from '@module/module-client-angular';
-
-const modules = new ClientModule(counter);
-modules.triggerOnCreate();
-
-export default modules;
