+++ tsconfig.json
@@ -20,6 +20,7 @@
   },
   "include": [
     "**/*.ts",
+    "**/*.d.ts",
     "**/*.tsx"
   ]
 }
